Rename utils.formatting to utils.code_repr
