+++ plasmapy/particles/tests/test_particle_class.py
@@ -30,7 +30,7 @@ from plasmapy.particles.particle_class import (
 )
 from plasmapy.particles.special_particles import ParticleZoo
 from plasmapy.utils import roman
-from plasmapy.utils.formatting import call_string
+from plasmapy.utils.code_repr import call_string
 from plasmapy.utils.pytest_helpers import run_test_equivalent_calls
 
 # (arg, kwargs, results_dict)
