+++ plasmapy/plasma/sources/plasmablob.py
@@ -11,7 +11,7 @@ from plasmapy.formulary.dimensionless import quantum_theta
 from plasmapy.formulary.parameters import _grab_charge
 from plasmapy.particles import particle_mass
 from plasmapy.plasma.plasma_base import GenericPlasma
-from plasmapy.utils import formatting
+from plasmapy.utils import code_repr
 from plasmapy.utils.decorators import validate_quantities
 from plasmapy.utils.exceptions import CouplingWarning
 
@@ -70,7 +70,7 @@ class PlasmaBlob(GenericPlasma):
             "Z": self.Z,
         }
 
-        return formatting.call_string(PlasmaBlob, (), argument_dict)
+        return code_repr.call_string(PlasmaBlob, (), argument_dict)
 
     @property
     def electron_temperature(self):
