+++ plasmapy/utils/pytest_helpers/tests/test_pytest_helpers.py
@@ -13,7 +13,7 @@ from plasmapy.tests.helpers import (
     UnexpectedResultFail,
 )
 from plasmapy.utils.exceptions import PlasmaPyError, PlasmaPyWarning
-from plasmapy.utils.formatting import call_string
+from plasmapy.utils.code_repr import call_string
 from plasmapy.utils.pytest_helpers import run_test, run_test_equivalent_calls
